scripts: live-wallpaper.sh early exit

+++ .local/share/scripts/live-wallpaper.sh
@@ -4,7 +4,9 @@ wallpapersdir="$HOME/.local/share/wallpapers/Live"
 
 wallpaper=$( ls "$wallpapersdir" | wofi --dmenu --prompt "🖼️" )
 
-wallpaper_path="$wallpapersdir/$type/$wallpaper"
+[ -z "$wallpaper" ] && exit 1
+
+wallpaper_path="$wallpapersdir/$wallpaper"
 
 pkill mpvpaper
 mpvpaper -s -o "no-audio --loop --panscan=1" ALL "$wallpaper_path"
